Delete copied over code
Since this is a rewirte, I thought it might be better to rewrite
the logic from scratch. This new code makes extensive use of GOTO
and switch statements, but I think it's more efficient and isn't
too hard to follow.

Again, none of this code has been tested yet. It probably doesn't
even compile.